test: add load testing with locust (#77)
* feat: locust for load tests

* wip

* perf: conn pool for sqlalchemy increased, password funcs are async

* fix: prohibit fav playlist deletion

* feat: tests done

* fix: track file not found

* feat: list with name search

+++ backend/services/accounts.py
@@ -1,6 +1,3 @@
-from datetime import datetime, timedelta
-from jose import jwt, JWTError
-from passlib.context import CryptContext
 from dto.accounts import (
     User,
     UserID,
@@ -33,6 +30,12 @@ from repositories.interfaces import (
     ITrackRepository,
 )
 from exceptions.music import ImageFileNotFoundException
+from exceptions.accounts import PlaylistFavDeletion
+
+from datetime import datetime, timedelta
+from jose import jwt, JWTError
+from passlib.context import CryptContext
+from starlette.concurrency import run_in_threadpool
 from typing import Optional
 from configs import environment as env
 
@@ -67,7 +70,7 @@ class AccountService:
         image_data: bytes | None = None,
         image_content_type: str | None = None,
     ) -> User:
-        hashed_password = self.get_hashed_password(new_user.password)
+        hashed_password = await self.get_hashed_password(new_user.password)
         user_with_hashed = new_user.model_copy(update={"password": hashed_password})
         user = await self.user_repository.create_user(user_with_hashed)
         if image_content_type:
@@ -123,7 +126,13 @@ class AccountService:
             PlaylistSearchParams(author_id=user_id.id, limit=10000)
         )
         for playlist in playlists:
-            await self.delete_playlist(PlaylistID(id=playlist.id))
+            await self.playlist_repository.delete_playlist(PlaylistID(id=playlist.id))
+            try:
+                await self.music_file_repository.delete_image(
+                    PlaylistID(id=playlist.id)
+                )
+            except ImageFileNotFoundException:
+                pass
 
         albums = await self.album_repository.get_albums(
             AlbumSearchParams(artist_id=user_id.id, limit=10000)
@@ -174,11 +183,15 @@ class AccountService:
         except JWTError:
             return None
 
-    def verify_password(self, entered_password: str, hashed_password: str) -> bool:
-        return self.pwd_context.verify(entered_password, hashed_password)
+    async def verify_password(
+        self, entered_password: str, hashed_password: str
+    ) -> bool:
+        return await run_in_threadpool(
+            self.pwd_context.verify, entered_password, hashed_password
+        )
 
-    def get_hashed_password(self, entered_password: str) -> str:
-        return self.pwd_context.hash(entered_password)
+    async def get_hashed_password(self, entered_password: str) -> str:
+        return await run_in_threadpool(self.pwd_context.hash, entered_password)
 
     # === Subscription-related methods ===
 
@@ -239,6 +252,9 @@ class AccountService:
         )
 
     async def delete_playlist(self, playlist_id: PlaylistID) -> None:
+        playlist = await self.playlist_repository.get_playlist_by_id(playlist_id)
+        if playlist.name == "fav":
+            raise PlaylistFavDeletion("Deletion of favorite playlist")
         await self.playlist_repository.delete_playlist(playlist_id)
         try:
             await self.music_file_repository.delete_image(playlist_id)
